requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
cvxopt==1.2.5.post1
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
idna==2.10
protobuf==3.13.0
psycopg2==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
python-decouple==3.4
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.11
whitenoise==5.2.0
